requirements.txt
anthropic>=0.50.0
pandas>=2.0.0
python-dotenv>=1.0.0
rich>=13.0.0
pytest>=8.0.0
